+++ libraries/template-haskell/changelog.md
@@ -10,6 +10,9 @@
 
   * Add support for unboxed sums. (#12478)
 
+  * Add support for attaching deriving strategies to `deriving` statements
+    (#10598)
+
 ## 2.11.0.0  *May 2016*
 
   * Bundled with GHC 8.0.1